add kafka storage tips

+++ deploy.sh
@@ -1,5 +1,12 @@
 #!/bin/bash
 
+if [[ ! -d ".deploy_git" ]]; then
+    echo "please clone blog to .deploy_git"
+    echo "git clone -b gh-page git@github.com:bwangelme/bwangelme.github.io .deploy_git"
+    exit 0
+fi
+
+
 echo "Reset Commit"
 cd .deploy_git
 git reset --hard origin/gh-page
